multiplied vitamins depending on serving size input along with allowing user to enter dosage amount for vitamins
Serving size code is on line 66 to allow for whatever the food item is to multiply all values of that food depending on what the user enters for serving size.
Rest of the code is for when a vitamin is searched up on the search bar, we now allow user to input a dosage amount of that vitamin whether it be in milligrams or grams.